Make `ExecutionContext` return `Halted` as soon as the program would halt.
Previously, we would not return `ExecutionStatus::Halted` until after
an additional step after the last instruction was executed. For example,
`+-><` would return `Running` for four steps, and then only on the fifth
step would it return `Halted`. This changes it so that `Halted` is instead
returned on the fourth step.